+++ aimmo/avatar_creator.py
@@ -1,5 +1,7 @@
+from aimmo.models import Avatar, Game, generate_auth_token
 
-from aimmo.models import Avatar, Game
+NUM_BYTES_FOR_TOKEN_GENERATOR = 16
+TOKEN_MAX_LENGTH = 24
 
 
 def create_avatar_for_user(user, game_id):
@@ -14,4 +16,5 @@ def create_avatar_for_user(user, game_id):
     game: Game = Game.objects.get(id=game_id)
     initial_code = game.worksheet.starter_code
     avatar = Avatar.objects.create(owner=user, code=initial_code, game_id=game_id)
+    avatar.auth_token = generate_auth_token(NUM_BYTES_FOR_TOKEN_GENERATOR, TOKEN_MAX_LENGTH)
     return avatar