+++ aimmo/game_creator.py
@@ -1,15 +1,10 @@
-import secrets
-
 from aimmo.avatar_creator import create_avatar_for_user
 from aimmo.game_manager import GameManager
+from aimmo.models import generate_auth_token
 from aimmo.serializers import GameSerializer
 
 NUM_BYTES_FOR_TOKEN_GENERATOR = 32
 TOKEN_MAX_LENGTH = 48
-
-
-def generate_game_token():
-    return secrets.token_urlsafe(nbytes=NUM_BYTES_FOR_TOKEN_GENERATOR)[:TOKEN_MAX_LENGTH]
 
 
 def create_game(main_user, form):
@@ -26,7 +21,7 @@ def create_game(main_user, form):
     :return: The initialised Game object.
     """
     game = form.save(commit=False)
-    game.auth_token = generate_game_token()
+    game.auth_token = generate_auth_token(NUM_BYTES_FOR_TOKEN_GENERATOR, TOKEN_MAX_LENGTH)
     game.generator = "Main"
     game.owner = game.game_class.teacher.new_user
     game.main_user = game.game_class.teacher.new_user