+++ aimmo_runner/minikube.py
@@ -2,6 +2,7 @@
 import atexit
 import os
 import platform
+import subprocess
 
 import kubernetes
 import yaml
@@ -49,7 +50,12 @@ def restart_pods():
     """
     print("Restarting pods")
 
-    run_command(["kubectl", "create", "-f", "agones/fleet.yml"])
+    try:
+        run_command(["kubectl", "create", "-f", "agones/fleet.yml"])
+    except subprocess.CalledProcessError:
+        run_command("kubectl delete fleet aimmo-game --ignore-not-found".split(" "))
+        run_command("kubectl delete --all deployment -n default".split(" "))
+        run_command(["kubectl", "create", "-f", "agones/fleet.yml"])
 
 
 def create_roles():