+++ game_frontend/src/containers/IDEEditor/index.test.js
@@ -40,6 +40,55 @@ describe('<IDEEditor />', () => {
     component.find('#post-code-button').simulate('click')
     expect(postCode).not.toBeCalled()
   })
+
+  it('resetCode called after clicking reset code', () => {
+    const props = {
+      resetCode: jest.fn(),
+      getCode: jest.fn(),
+    }
+    window.confirm = jest.fn(() => true)
+
+    const component = createShallowWithTheme(<IDEEditor {...props} />, 'dark')
+
+    component.find('#reset-code-button').simulate('click')
+    expect(props.resetCode).toBeCalled()
+  })
+
+  it('togglePauseGame called after clicking pause', () => {
+    const props = {
+      getCode: jest.fn(),
+      togglePauseGame: jest.fn()
+    }
+
+    const component = createShallowWithTheme(<IDEEditor {...props} />, 'dark')
+
+    component.find('#game-pause-button').simulate('click')
+    expect(props.togglePauseGame).toBeCalled()
+  })
+
+  it('gameResumed called after clicking post code', () => {
+    const props = {
+      getCode: jest.fn(),
+      postCode: jest.fn(),
+      gameResume: jest.fn()
+    }
+
+    const component = createShallowWithTheme(<IDEEditor {...props} />, 'dark')
+
+    component.instance().postCode();
+    expect(props.gameResume).toBeCalled()
+  })
+
+  it('gameResumed called after clicking post code', () => {
+    const props = {
+      getCode: jest.fn(),
+    }
+
+    const component = createShallowWithTheme(<IDEEditor {...props} />, 'dark')
+
+    component.instance().codeChanged("someCode");
+    expect(component.state('code')).toEqual("someCode");
+  })
 })
 
 describe('<IDEEditorLayout />', () => {