+++ game_frontend/src/pyodide/webWorker.test.ts
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-import { simplifyErrorMessageInLog } from './webWorker'
+import { computeNextAction, simplifyErrorMessageInLog } from './webWorker'
 jest.mock('threads/worker')
 
 describe('Error formatting', () => {
@@ -18,5 +18,45 @@ Uh oh! Something isn't correct on line 3. Here's the error we got:
 AttributeError: module 'simulation.direction' has no attribute 'NRTH'
 `.trim()
     expect(result).toBe(expected)
+  })
+})
+
+describe('pyodide webWorker', () => {
+  it('returns wait action if game is paused', async () => {
+    expect.assertions(1);
+    const isGamePaused = true
+    const playerAvatarID = 1
+    const turnCount = 5
+    const gameState = {
+      "players": [
+        {
+          "location": {
+            "x": -13,
+            "y": 13
+          },
+          "id": 1,
+          "orientation": "north"
+        },
+        {
+          "location": {
+            "x": 6,
+            "y": 10
+          },
+          "id": 2,
+          "orientation": "north"
+        }
+      ],
+      "turnCount": turnCount,
+    }
+
+    const expected = {
+      action: {
+        action_type: "wait",
+      },
+      log: '',
+      turnCount: turnCount + 1,
+    }
+
+    return expect(computeNextAction(gameState, playerAvatarID, isGamePaused)).resolves.toEqual(expected);
   })
 })