+++ game_frontend/src/redux/features/AvatarWorker/epics.js
@@ -74,7 +74,8 @@ const computeNextActionEpic = (
         switchMap(({ payload: { gameState } }) =>
           computeNextAction$(
             gameState,
-            state$.value.game.connectionParameters.currentAvatarID
+            state$.value.game.connectionParameters.currentAvatarID,
+            state$.value.game.gamePaused,
           ).pipe(timeoutIfWorkerTakesTooLong(state$, resetWorker, scheduler))
         ),
         tap(socket.emitAction),
@@ -148,6 +149,7 @@ const checkBadgesEarnedEpic = (action$, state$, { pyodideRunner: { checkIfBadgeE
               computedTurnResult,
               state$.value.editor.code.codeOnServer,
               state$.value.game.gameState
+              // state$.value.game.connectionParameters.currentAvatarID
             )
           )
         ),