+++ game_frontend/src/redux/features/Game/actions.js
@@ -33,6 +33,14 @@ const gameLoaded = () => ({
   type: types.GAME_LOADED,
 })
 
+const togglePauseGame = () => ({
+  type: types.TOGGLE_PAUSE_GAME,
+})
+
+const gameResume = () => ({
+  type: types.GAME_RESUME,
+})
+
 const setTimeout = () => ({
   type: types.SET_TIMEOUT,
 })
@@ -55,4 +63,6 @@ export default {
   setTimeout,
   mapPanned,
   centerCameraOnUserAvatar,
+  togglePauseGame,
+  gameResume,
 }