+++ game_frontend/src/redux/features/Game/reducers.js
@@ -26,6 +26,16 @@ const gameReducer = (state = {}, action) => {
         ...state,
         cameraCenteredOnUserAvatar: true,
       }
+    case types.TOGGLE_PAUSE_GAME:
+      return {
+        ...state,
+        gamePaused: !state.gamePaused,
+      }
+    case types.GAME_RESUME:
+      return {
+        ...state,
+        gamePaused: false,
+      }
     default:
       return state
   }